Moving the rest of file tests over to File
