+++ radio/nsl-duplex-d2/src/lib.rs
@@ -173,50 +173,4 @@ mod tests {
         let data: Vec<u8> = Vec::new();
         assert!(d.send(data).is_ok(), "Send should pass")
     }
-
-    #[test]
-    fn test_uploaded_file_count_one() {
-        let mut ret_msg = Vec::<u8>::new();
-        ret_msg.extend(RESP_HEADER.as_bytes().iter().cloned());
-        ret_msg.push(1 as u8);
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        let d = DuplexD2 {
-            conn: Box::new(TestConnection { data: ret_msg }),
-        };
-        let count = d.get_uploaded_file_count().unwrap();
-        assert_eq!(count, 1, "File count should be one")
-    }
-
-    #[test]
-    fn test_uploaded_file_count_zero() {
-        let mut ret_msg = Vec::<u8>::new();
-        ret_msg.extend(RESP_HEADER.as_bytes().iter().cloned());
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        let d = DuplexD2 {
-            conn: Box::new(TestConnection { data: ret_msg }),
-        };
-        let count = d.get_uploaded_file_count().unwrap();
-        assert_eq!(count, 0, "File count should be zero")
-    }
-
-    #[test]
-    fn test_uploaded_file_count_many() {
-        let mut ret_msg = Vec::<u8>::new();
-        ret_msg.extend(RESP_HEADER.as_bytes().iter().cloned());
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        ret_msg.push(0 as u8);
-        ret_msg.push(1 as u8);
-        let d = DuplexD2 {
-            conn: Box::new(TestConnection { data: ret_msg }),
-        };
-        let count = d.get_uploaded_file_count().unwrap();
-        assert_eq!(count, 16777216, "File count should be 16777216")
-    }
-
 }
